Do not output varbit data by default
The value can be large, and for large tables, PostgreSQL will accumulate
the tuples. A 100GB table could have burn 25 MB of shared buffers, at least.
Even if the user won't use the map, but just aggregated data.

It is possible to get the varbit by adding , a 'true' as last parameter of
pgfincore()