Add pg_sysconf_size, vm_free_size, drop pgsysconf
pg_sysconf_size returns 'immutable' sizes from sysconf and pg:
- postgres page size
- system page size
- total memory
vm_free_size returns the current free memory on the system

Remove pgsysconf datum which is obsolete, and add an SQL wrapper on top
of the new functions.
Users are encouraged to update their code using pgsysconf which is now
deprecated and will be removed in the future.