Start deprecation of PostgreSQL < 12
11 is EOL since 9/11/2023 and the ereport() functions has changed in
PostgreSQL 12. Also it's time to modernize code and SMGRelation was
absent in 9.6... so for the new code it'll be easier to manage only
current PostgreSQL versions.

And prepare for pgfincore 2.0.0 !

As we're to deprecated some C functions, we need an upgrade script from
1.3.1 but also a fresh new one for 2.0.0 else PostgreSQL will fail to
apply upgrade path (first try to create pgsysconf will fail for example,
as the shared object don't have this function anymore.