Add linux 6.5 cachestat syscall
Linux 6.5 introduced sys_cachestat:
https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/mm/filemap.c?h=v6.5#n4174

This commit adds a vm_relation_cachestat() which allows to execute cachestat() on
all or a part of a relation, it returns by default a row per segment of the
relation.

range parameter is usefull to go into the details, and offset/length to
join with pg_buffercache for example.